# Allow small matrix values introduced by presolve, also in debug mode