@jajhall 2023-02-26
Here the assert is important, as very small matrix values must be avoided. In the LU factorization, it is assumed that any structural nonzero can be used as a pivot. Rather than just accept this PR, I'll add some code to analyse the small values and do something about them.