@ambros-gleixner 2023-02-27
Thanks. In debug mode the entries seem to be deleted already, in HighsMatrixUtils.cpp:204, but not in optimized mode I guess?

So currently there is also a difference between debug and optimized behavior that can make debugging harder. I believe this difference should be eliminated independently of / in addition to fixing the existence of these small entries.