@jajhall 2023-03-01
The difference isn't so bad: if any matrix values are changed in debug mode, the assert will be triggered since `assessLp` will return a warning, so debug and release will never run differently. For this reason I won't make this PR, since debug and release would diverge as a result