@ambros-gleixner 2023-03-01
True, the current PR would not be good in that respect, true.

But there is still a problem here. The aggregator introduces matrix coefficients below small_matrix_threshold although the original problem does not contain them. If I understand correctly, these are not removed in optimized mode, right?

This can also be reproduced with HiGHS itself on the command line, running 
[debug.lp.txt](https://github.com/ERGO-Code/HiGHS/files/10864362/debug.lp.txt) on latest:

```
Running HiGHS 1.5.1 [date: 2023-03-01, git hash: 355fb597d]
Copyright (c) 2023 HiGHS under MIT licence terms
LP   debug has 13499 rows; 2712 cols; 38172 nonzeros
Presolving model
414 rows, 407 cols, 2613 nonzeros
247 rows, 241 cols, 2306 nonzeros
222 rows, 222 cols, 2117 nonzeros
WARNING: LP matrix packed vector contains 18 |values| in [3.98097e-13, 9.99975e-10] less than or equal to 1e-09: ignored
highs: /home/optimi/bzfgleix/opti/devel/HiGHS/src/lp_data/Highs.cpp:1134: HighsStatus Highs::run(): Assertion `assessLp(reduced_lp, options_) == HighsStatus::kOk' failed.
Abort
```

I can file a proper issue if you want to.