@ambros-gleixner 2023-03-02
Then I am confused, you said 
> ... very small matrix values must be avoided. In the LU factorization, it is assumed that any structural nonzero can be used as a pivot.

:)