@jajhall 2023-03-02
Yes, this can happen in theory. However, it's clearly very unlikely since HiGHS in release solves so many of the MIPLIB 2017 test problems. Pivoting on a small value would probably imply singularity of the basis matrix and, of course, simplex has safeguards to avoid this.