@ambros-gleixner 2023-03-03
If you consider the optimized behavior okay, then I think it should not throw an assert in debug mode. Otherwise we miss out on the chance of testing the code path after the assert in debug mode.

Modifying the assert as I did is of course bad as discussed above.

I would then suggest to have a version of assessLp() that is not intrusive, but only checks the matrix and prints warnings. And with `kAllowDeveloperAssert=true` it could turn on the current behavior of stopping.